Code cleanup

+++ source/RKMisc.c
@@ -79,7 +79,7 @@ char *RKIntegerToCommaStyleString(const long num) {
 //
 char *RKFloatToCommaStyleString(const double num) {
     char *intString = RKIntegerToCommaStyleString((long)num);
-    snprintf(intString + strlen(intString), 32 - strlen(intString), ".%03.0f", 1000.0f * (num - floorf(num))); 
+    snprintf(intString + strlen(intString), 32 - strlen(intString), ".%03.0f", 1000.0f * (num - floor(num))); 
     return intString;
 }
